test(web): split ingestion api server tests
